pyproject.toml
[project]
name = "lecturetube-recorder"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "pytz>=2025.1",
    "pyyaml>=6.0.2",
    "schedule>=1.2.2",
]
